Make interpolation script have O(1) space complexity for d dimensions (previously O(2^d) and not configurable to arbitrarily many), combine interpolation binary counter loop with main interpolation one
Two commits in one, the first didn't seem to be committed before I made the second.
I want to add orthographic projection of each pixel's number of neighbours exceeded in 3D (hiding those that exceed exactly 13 to reveal the network of disparities).
After making the two loops and noticing their conditions were very similar, with one having a different indice beginning and checking for the opposite, the revelation came to me that they could be combined, each dimension being turned to 0 (it counts up the first element and carries the one in-place in front of the interpolator).
I began the list indice at 1 in the optimised do interpolation function because it needs to check each new dimension's counter before interpolating so the increment is at the end, but it would be 0 like usual in a Python script.
Anyway, look at the do all interpolation function, it achieves in 18 lines (15 if there were an exponentiation function, but in this case a `for i in` loop wouldn't help because it's all conditional) what previously took O(2^(d+1)-1) lines, you will cry. Many combinations were tried and reverted (that gave strange square artefacts) before the correct one.