Create Python version (working but no graphics) in n dimensions with beautiful interpolation algorithm
(I discovered the algorithm when making the Scratch version, where the axes it interpolates in are based on the digits through which the 1 is carried in the corresponding iteration of binary counting, and the simple boolean algebra reduces this to a single while loop instead of two.)
Is entirely the same logic as the Scratch one, but working in an unbounded number of dimensions (though I can only generate unit vectors in three currently (by converting a random location in the cylindrical equi-area projection (which has an even distribution (this results from (and proves) the fact that if you only look at the x positions of evenly-distributed points on a sphere's surface, they're themselves evenly distributed (because the ones towards the top and bottom are in thinner bands than in the equator but that are closer to being perpendicular to the x axis (and thus more densely-distributed) and these cancel out))))).
I will next add graphics (initially with slices like an MRI, but eventually with quaternion rotations and such) (like the Scratch one).